Release 1.3.0 (beta)

+++ app/src/main/java/pl/temomuko/autostoprace/ui/base/drawer/DrawerActivity.java
@@ -89,7 +89,7 @@ public abstract class DrawerActivity extends BaseActivity implements DrawerMvpVi
     }
 
     @Override
-    public void setupTeamNumberText(int teamNumber) {
+    public void setupTeamNumberText(long teamNumber) {
         ((LinearLayout) mTeamTextView.getParent()).setVisibility(View.VISIBLE);
         mTeamTextView.setText(String.valueOf(teamNumber));
     }
